client back end connect

+++ server/src/recipes/recipe.controller.ts
@@ -3,25 +3,37 @@ import {
   Controller,
   Delete,
   Get,
+  Headers,
   Param,
   Patch,
   Post,
   Put,
   Query,
   UploadedFiles,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { RecipeService } from './recipes.service';
 import { RecipeDto } from './dto/recipe.dto';
 import { FileFieldsInterceptor } from '@nestjs/platform-express';
+import { JwtService } from '@nestjs/jwt';
+import { CreateUserDto } from 'src/users/dto/user.create.dto';
+import { CheckRoleGuard } from 'src/role/role.guard';
+import { CheckRole } from 'src/role/role.decorator';
+import { ObjectId } from 'mongodb';
+import { UserService } from 'src/users/users.service';
 
-@Controller('recipes')
+@Controller('recipe')
 export class RecipeController {
-  constructor(private readonly recipeService: RecipeService) {}
+  constructor(
+    private readonly recipeService: RecipeService,
+    private readonly jwtService: JwtService,
+    private readonly userService: UserService,
+  ) {}
 
-  @Post('add')
+  @Post('create')
   @UseInterceptors(FileFieldsInterceptor([{ name: 'images' }]))
-  async addRecipe(
+  async createRecipe(
     @Body() body: { body: string },
     @UploadedFiles()
     files?: {
@@ -36,7 +48,7 @@ export class RecipeController {
       );
       req.images.push(...url);
     }
-    return this.recipeService.addRecipe(req);
+    return this.recipeService.createRecipe(req);
   }
 
   @Patch('upload/:id')
@@ -44,53 +56,100 @@ export class RecipeController {
   async uploadImage(
     @Param('id') id: string,
     @Body() body: { body: string },
+    @Headers() Headers: any,
     @UploadedFiles()
     files?: {
       images?: Express.Multer.File[];
     },
   ) {
-    const req: RecipeDto = JSON.parse(body.body);
+    try {
+      const [type, token] = Headers.authorization?.split(' ') ?? [];
+      if (!token) {
+        return 'have not token';
+      }
+      const decodedToken: any = this.jwtService.decode(token);
+      if (!decodedToken) {
+        return 'token extist';
+      }
+      const recipe: any = await this.recipeService.getRecipe(id);
+      const req: RecipeDto = JSON.parse(body.body);
+      if (decodedToken.role == true || decodedToken._id == recipe.created_by) {
+        if (files?.images) {
+          const url = await this.recipeService.uploadImageToCloudinary(
+            files.images,
+          );
+          req.images.push(...url);
+        }
 
-    if (files?.images) {
-      const url = await this.recipeService.uploadImageToCloudinary(
-        files.images,
-      );
-      req.images.push(...url);
+        return this.recipeService.editRecipe(id, req);
+      } else {
+        return 'you have not permission';
+      }
+    } catch (error) {
+      return { error };
     }
-    req.status = 'pending';
-
-    return this.recipeService.editRecipe(id, req);
-  }
-  @Put('approve')
-  recipeApprove(@Body('id') id: string) {
-    return this.recipeService.recipeApprove(id);
-  }
-  @Put('deny')
-  recipeDeny(@Body('id') id: string) {
-    return this.recipeService.recipeDeny(id);
-  }
-  @Get('user/:id')
-  getUserRecipe(@Param('id') id: string) {
-    const result = this.recipeService.getUserRecipe(id);
-    return result;
-  }
-  @Get('status')
-  getStatus() {
-    return this.recipeService.getStatus();
   }
 
-  @Get('all')
-  getRecipes() {
-    return this.recipeService.getRecipes();
+  @Put('status/:id')
+  @UseGuards(CheckRoleGuard)
+  @CheckRole(true)
+  recipeStatus(@Param('id') id: string, @Body('status') status: string) {
+    return this.recipeService.recipeStatus(id, status);
   }
 
-  @Get('pending')
-  getPendingRecipes() {
-    return this.recipeService.getPendingRecipes();
+  @Get('statistics')
+  getStatistics() {
+    return this.recipeService.getStatistics();
   }
-  @Get('filter')
-  getFilterRecipes(@Query() filter: any) {
-    return this.recipeService.getFilterRecipe(filter);
+
+  @Get('recipes')
+  async getRecipes(@Query() query: any) {
+    let optionQuery: any = { $and: [] };
+
+    if (query.status) {
+      optionQuery.$and.push({ status: query.status });
+    }
+    if (query.category) {
+      const categoryIds = Array.isArray(query.category)
+        ? query.category.map((categoryId: string) => new ObjectId(categoryId))
+        : [new ObjectId(query.category)];
+
+      optionQuery.$and.push({ categories: { $all: categoryIds } });
+    }
+    if (query.tag) {
+      const tagIds = Array.isArray(query.tag)
+        ? query.tag.map((tagId: string) => new ObjectId(tagId))
+        : [new ObjectId(query.tag)];
+
+      optionQuery.$and.push({ tags: { $all: tagIds } });
+    }
+    if (query.ingredient) {
+      const ingredientNames = Array.isArray(query.ingredient)
+        ? query.ingredient.map((ingredient: string) => ingredient)
+        : [query.ingredient];
+      console.log(ingredientNames, 'this is worked');
+      optionQuery.$and.push({ ingredients: { $all: ingredientNames } });
+    }
+    if (query.user) {
+      const userId = new ObjectId(query.user);
+      optionQuery.$and.push({ created_by: userId });
+    }
+    if (query.favorites) {
+      const userfavorites = await this.userService.getUser(query.favorites);
+      console.log(userfavorites);
+      const favoriteIds = userfavorites.favorites.map(
+        (Id: string) => new ObjectId(Id),
+      );
+
+      optionQuery.$and.push({ _id: { $in: favoriteIds } });
+    }
+    if (Object.keys(query).length == 0) {
+      optionQuery = {};
+    }
+    console.log(query);
+    console.log(optionQuery);
+
+    return this.recipeService.getRecipes(optionQuery);
   }
 
   @Get(':id')
@@ -98,13 +157,26 @@ export class RecipeController {
     return this.recipeService.getRecipe(id);
   }
 
-  @Put(':id')
-  editRecipe(@Param('id') id: string, @Body() recipeDto: RecipeDto) {
-    return this.recipeService.editRecipe(id, recipeDto);
-  }
-
   @Delete(':id')
-  deleteRecipe(@Param('id') id: string) {
-    return this.recipeService.deleteRecipe(id);
+  async deleteRecipe(@Param('id') id: string, @Headers() Headers) {
+    try {
+      const [type, token] = Headers.authorization?.split(' ') ?? [];
+
+      if (!token) {
+        return 'have not token';
+      }
+      const decodedToken: any = this.jwtService.decode(token);
+      if (!decodedToken) {
+        return 'token extist';
+      }
+      const recipe: any = await this.recipeService.getRecipe(id);
+      if (decodedToken.role == true || decodedToken._id == recipe.created_by) {
+        return this.recipeService.deleteRecipe(id);
+      } else {
+        return 'you have not permission';
+      }
+    } catch (error) {
+      return error;
+    }
   }
 }
